move confirm to util package
